// Back/src/schemas/types/book.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface Book {
  book_id?: number;
  author_id: number;
  title: string;
  publisher_id: number;
  place_publication: string;
  published_dt: number;
  category_id: number;
  isbn: number;
  status: boolean;
}